CE-2531: Added Special:ContentReview subpage with archive data for given wiki
