+++ extensions/wikia/ContentReview/models/ReviewModel.php
@@ -150,39 +150,6 @@ class ReviewModel extends ContentReviewBaseModel {
 		return true;
 	}
 
-	/**
-	 * Backup completed review in log table
-	 *
-	 * @param Array $review
-	 * @param int $status
-	 * @return bool
-	 * @throws \FluentSql\Exception\SqlException
-	 */
-	public function backupCompletedReview( $review, $status, $reviewUserId ) {
-		$db = $this->getDatabaseForWrite();
-
-		( new \WikiaSQL() )
-			->INSERT( self::CONTENT_REVIEW_LOG_TABLE )
-			->SET( 'wiki_id', $review['wiki_id'] )
-			->SET( 'page_id', $review['page_id'] )
-			->SET( 'revision_id', $review['revision_id'] )
-			->SET( 'status', $status )
-			->SET( 'submit_user_id', $review['submit_user_id'] )
-			->SET( 'submit_time', $review['submit_time'] )
-			->SET( 'review_user_id', $reviewUserId )
-			->SET( 'review_start', $review['review_start'] )
-			// review_end has a default value set to CURRENT_TIMESTAMP
-			->run( $db );
-
-		$affectedRows = $db->affectedRows();
-
-		if ( $affectedRows === 0 ) {
-			throw new \FluentSql\Exception\SqlException( 'The INSERT operation failed.' );
-		}
-
-		return true;
-	}
-
 	public function updateRevisionStatus( $wiki_id, $page_id, $oldStatus, $status, $reviewerId  ) {
 		$db = $this->getDatabaseForWrite();
